@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
-migration_class = ActiveRecord::VERSION::MAJOR >= 5 ? ActiveRecord::Migration[4.2] : ActiveRecord::Migration
-class CreatePublicTokens < migration_class
+class CreatePublicTokens < ActiveRecord::Migration[4.2]
   def up
     create_table :public_tokens do |t|
       t.string :token
