@@ -5,16 +5,15 @@ require 'active_support/core_ext/object/blank'
 require 'forwardable'
 require 'active_record'
 require 'apartment/tenant'
+require 'apartment/deprecation'
 
 require_relative 'apartment/log_subscriber'
+require_relative 'apartment/active_record/connection_handling'
+require_relative 'apartment/active_record/schema_migration'
+require_relative 'apartment/active_record/internal_metadata'
 
-if ActiveRecord.version.release >= Gem::Version.new('6.0')
-  require_relative 'apartment/active_record/connection_handling'
-end
-
-if ActiveRecord.version.release >= Gem::Version.new('6.1')
-  require_relative 'apartment/active_record/schema_migration'
-  require_relative 'apartment/active_record/internal_metadata'
+if ActiveRecord.version.release >= Gem::Version.new('7.1')
+  require_relative 'apartment/active_record/postgres/schema_dumper'
 end
 
 # Apartment main definitions
@@ -23,7 +22,8 @@ module Apartment
     extend Forwardable
 
     ACCESSOR_METHODS = %i[use_schemas use_sql seed_after_create prepend_environment default_tenant
-                          append_environment with_multi_server_setup tenant_presence_check active_record_log].freeze
+                          append_environment with_multi_server_setup tenant_presence_check
+                          active_record_log pg_exclude_clone_tables].freeze
 
     WRITER_METHODS = %i[tenant_names database_schema_file excluded_models
                         persistent_schemas connection_class
@@ -33,14 +33,10 @@ module Apartment
     attr_accessor(*ACCESSOR_METHODS)
     attr_writer(*WRITER_METHODS)
 
-    if ActiveRecord.version.release >= Gem::Version.new('6.1')
-      def_delegators :connection_class, :connection, :connection_db_config, :establish_connection
+    def_delegators :connection_class, :connection, :connection_db_config, :establish_connection
 
-      def connection_config
-        connection_db_config.configuration_hash
-      end
-    else
-      def_delegators :connection_class, :connection, :connection_config, :establish_connection
+    def connection_config
+      connection_db_config.configuration_hash
     end
 
     # configure apartment with available options
@@ -57,7 +53,7 @@ module Apartment
     end
 
     def tld_length=(_)
-      Apartment::Deprecation.warn('`config.tld_length` have no effect because it was removed in https://github.com/influitive/apartment/pull/309')
+      Apartment::DEPRECATOR.warn('`config.tld_length` have no effect because it was removed in https://github.com/influitive/apartment/pull/309')
     end
 
     def db_config_for(tenant)
