@@ -13,8 +13,8 @@ module Apartment
 
         migration_scope_block = ->(migration) { ENV['SCOPE'].blank? || (ENV['SCOPE'] == migration.scope) }
 
-        if activerecord_below_5_2?
-          ActiveRecord::Migrator.migrate(ActiveRecord::Migrator.migrations_paths, version, &migration_scope_block)
+        if ActiveRecord.version >= Gem::Version.new('7.2.0')
+          ActiveRecord::Base.connection_pool.migration_context.migrate(version, &migration_scope_block)
         else
           ActiveRecord::Base.connection.migration_context.migrate(version, &migration_scope_block)
         end
@@ -24,8 +24,8 @@ module Apartment
     # Migrate up/down to a specific version
     def run(direction, database, version)
       Tenant.switch(database) do
-        if activerecord_below_5_2?
-          ActiveRecord::Migrator.run(direction, ActiveRecord::Migrator.migrations_paths, version)
+        if ActiveRecord.version >= Gem::Version.new('7.2.0')
+          ActiveRecord::Base.connection_pool.migration_context.run(direction, version)
         else
           ActiveRecord::Base.connection.migration_context.run(direction, version)
         end
@@ -35,18 +35,12 @@ module Apartment
     # rollback latest migration `step` number of times
     def rollback(database, step = 1)
       Tenant.switch(database) do
-        if activerecord_below_5_2?
-          ActiveRecord::Migrator.rollback(ActiveRecord::Migrator.migrations_paths, step)
+        if ActiveRecord.version >= Gem::Version.new('7.2.0')
+          ActiveRecord::Base.connection_pool.migration_context.rollback(step)
         else
           ActiveRecord::Base.connection.migration_context.rollback(step)
         end
       end
-    end
-
-    private
-
-    def activerecord_below_5_2?
-      ActiveRecord.version.release < Gem::Version.new('5.2.0')
     end
   end
 end
