@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-if defined?(JRUBY_VERSION)
+if defined?(JRUBY_VERSION) && ENV['DATABASE_ENGINE'] == 'mysql'
 
   require 'spec_helper'
   require 'apartment/adapters/jdbc_mysql_adapter'
@@ -9,8 +9,8 @@ if defined?(JRUBY_VERSION)
     subject(:adapter) { Apartment::Tenant.adapter }
 
     def tenant_names
-      ActiveRecord::Base.connection.execute('SELECT schema_name FROM information_schema.schemata').collect do |row|
-        row['schema_name']
+      ActiveRecord::Base.connection.execute('SELECT SCHEMA_NAME FROM information_schema.schemata').collect do |row|
+        row['SCHEMA_NAME']
       end
     end
 
