@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
-require 'spec_helper'
-require 'apartment/adapters/postgresql_adapter'
+if !defined?(JRUBY_VERSION) && ENV['DATABASE_ENGINE'] == 'postgresql'
 
-describe Apartment::Adapters::PostgresqlAdapter, database: :postgresql do
-  unless defined?(JRUBY_VERSION)
+  require 'spec_helper'
+  require 'apartment/adapters/postgresql_adapter'
 
+  describe Apartment::Adapters::PostgresqlAdapter, database: :postgresql do
     subject { Apartment::Tenant.adapter }
 
     it_behaves_like 'a generic apartment adapter callbacks'
@@ -62,6 +62,62 @@ describe Apartment::Adapters::PostgresqlAdapter, database: :postgresql do
       it_behaves_like 'a generic apartment adapter'
       it_behaves_like 'a generic apartment adapter able to handle custom configuration'
       it_behaves_like 'a connection based apartment adapter'
+    end
+
+    context 'when using pg_exclude_clone_tables with SQL dump' do
+      before do
+        Apartment.excluded_models = ['Company']
+        Apartment.use_schemas = true
+        Apartment.use_sql = true
+        Apartment.pg_exclude_clone_tables = true
+        ActiveRecord::Base.connection.execute <<-PROCEDURE
+          CREATE OR REPLACE FUNCTION test_function() RETURNS INTEGER AS $function$
+          DECLARE
+            r1 INTEGER;
+            r2 INTEGER;
+          BEGIN
+            SELECT COUNT(*) INTO r1 FROM public.companies;
+            SELECT COUNT(*) INTO r2 FROM public.users;
+            RETURN r1 + r2;
+          END;
+          $function$ LANGUAGE plpgsql;
+        PROCEDURE
+      end
+
+      after do
+        Apartment::Tenant.drop('has-procedure') if Apartment.connection.schema_exists? 'has-procedure'
+        ActiveRecord::Base.connection.execute('DROP FUNCTION IF EXISTS test_function();')
+        # Apartment::Tenant.init creates per model connection.
+        # Remove the connection after testing not to unintentionally keep the connection across tests.
+        Apartment.excluded_models.each do |excluded_model|
+          excluded_model.constantize.remove_connection
+        end
+      end
+
+      # Not sure why, but somehow using let(:tenant_names) memoizes for the whole example group, not just each test
+      def tenant_names
+        ActiveRecord::Base.connection.execute('SELECT nspname FROM pg_namespace;').collect { |row| row['nspname'] }
+      end
+
+      let(:default_tenant) { subject.switch { ActiveRecord::Base.connection.schema_search_path.delete('"') } }
+      let(:c) { rand(5) }
+      let(:u) { rand(5) }
+
+      it_behaves_like 'a generic apartment adapter'
+      it_behaves_like 'a schema based apartment adapter'
+
+      # rubocop:disable RSpec/ExampleLength
+      it 'not change excluded_models in the procedure code' do
+        Apartment::Tenant.init
+        Apartment::Tenant.create('has-procedure')
+        Apartment::Tenant.switch!('has-procedure')
+        c.times { Company.create }
+        u.times { User.create }
+        count = ActiveRecord::Base.connection.execute('SELECT test_function();')[0]['test_function']
+        expect(count).to(eq(Company.count + User.count))
+        Company.delete_all
+      end
+      # rubocop:enable RSpec/ExampleLength
     end
   end
 end
