@@ -23,7 +23,7 @@ Apartment.configure do |config|
   # use postgres schemas?
   config.use_schemas = true
 
-  # use raw SQL dumps for creating postgres schemas? (only appies with use_schemas set to true)
+  # use raw SQL dumps for creating postgres schemas? (only applies with use_schemas set to true)
   # config.use_sql = true
 
   # configure persistent schemas (E.g. hstore )
