@@ -63,7 +63,7 @@ shared_examples_for 'a schema based apartment adapter' do
   describe '#create' do
     it 'should load schema.rb to new schema' do
       connection.schema_search_path = schema1
-      expect(connection.tables).to include('companies')
+      expect(connection.tables).to include('users')
     end
 
     it 'should yield to block if passed and reset' do
@@ -92,6 +92,27 @@ shared_examples_for 'a schema based apartment adapter' do
       end
 
       after { subject.drop(db) }
+    end
+
+    context 'with a default_tenant', default_tenant: true do
+      let(:from_default_tenant) { 'new_from_custom_default_tenant' }
+
+      before do
+        subject.create(from_default_tenant)
+      end
+
+      after do
+        subject.drop(from_default_tenant)
+      end
+
+      it 'should correctly create the new schema' do
+        expect(tenant_names).to include(from_default_tenant)
+      end
+
+      it 'should load schema.rb to new schema' do
+        connection.schema_search_path = from_default_tenant
+        expect(connection.tables).to include('users')
+      end
     end
   end
 
@@ -128,9 +149,12 @@ shared_examples_for 'a schema based apartment adapter' do
       end
     end
 
-    # rubocop:disable RSpec/MultipleExpectations
     it 'connects and resets' do
       subject.switch(schema1) do
+        # Ensure sequence is not cached
+        Company.reset_sequence_name
+        User.reset_sequence_name
+
         expect(connection.schema_search_path).to start_with %("#{schema1}")
         expect(User.sequence_name).to eq "#{User.table_name}_id_seq"
         expect(Company.sequence_name).to eq "#{public_schema}.#{Company.table_name}_id_seq"
@@ -140,12 +164,29 @@ shared_examples_for 'a schema based apartment adapter' do
       expect(User.sequence_name).to eq "#{User.table_name}_id_seq"
       expect(Company.sequence_name).to eq "#{public_schema}.#{Company.table_name}_id_seq"
     end
-    # rubocop:enable RSpec/MultipleExpectations
 
-    it 'allows a list of schemas' do
-      subject.switch([schema1, schema2]) do
-        expect(connection.schema_search_path).to include %("#{schema1}")
-        expect(connection.schema_search_path).to include %("#{schema2}")
+    describe 'multiple schemas' do
+      it 'allows a list of schemas' do
+        subject.switch([schema1, schema2]) do
+          expect(connection.schema_search_path).to include %("#{schema1}")
+          expect(connection.schema_search_path).to include %("#{schema2}")
+        end
+      end
+
+      it 'connects and resets' do
+        subject.switch([schema1, schema2]) do
+          # Ensure sequence is not cached
+          Company.reset_sequence_name
+          User.reset_sequence_name
+
+          expect(connection.schema_search_path).to start_with %("#{schema1}")
+          expect(User.sequence_name).to eq "#{User.table_name}_id_seq"
+          expect(Company.sequence_name).to eq "#{public_schema}.#{Company.table_name}_id_seq"
+        end
+
+        expect(connection.schema_search_path).to start_with %("#{public_schema}")
+        expect(User.sequence_name).to eq "#{User.table_name}_id_seq"
+        expect(Company.sequence_name).to eq "#{public_schema}.#{Company.table_name}_id_seq"
       end
     end
   end
@@ -213,7 +254,7 @@ shared_examples_for 'a schema based apartment adapter' do
       it 'should not raise any errors' do
         expect do
           subject.switch! 'unknown_schema'
-        end.not_to raise_error(Apartment::TenantNotFound)
+        end.not_to raise_error
       end
     end
 
