@@ -5,9 +5,7 @@ require 'yaml'
 module Apartment
   module Test
     def self.config
-      # rubocop:disable Security/YAMLLoad
-      @config ||= YAML.load(ERB.new(IO.read('spec/config/database.yml')).result)
-      # rubocop:enable Security/YAMLLoad
+      @config ||= YAML.safe_load(ERB.new(File.read('spec/config/database.yml')).result)
     end
   end
 end
