@@ -17,7 +17,7 @@ module Apartment
             def config
               db = RSpec.current_example.metadata.fetch(:database, :postgresql)
 
-              Apartment::Test.config['connections'][db.to_s].symbolize_keys
+              Apartment::Test.config['connections'][db.to_s]&.symbolize_keys
             end
 
             # before
@@ -27,14 +27,13 @@ module Apartment
             example.run
 
             # after
-            Rails.configuration.database_configuration = {}
-            ActiveRecord::Base.clear_all_connections!
+            ActiveRecord::Base.connection_handler.clear_all_connections!
 
             Apartment.excluded_models.each do |model|
               klass = model.constantize
 
-              Apartment.connection_class.remove_connection(klass)
-              klass.clear_all_connections!
+              klass.remove_connection
+              klass.connection_handler.clear_all_connections!
               klass.reset_table_name
             end
             Apartment.reset
